adc: Change the values returned when reading the ADC to 16 bit
The maximum size of a read value is 16 bits wide, even though the register is 32
bits wide. The other 16 are always 0.

+++ src/adc.rs
@@ -213,8 +213,8 @@ impl<'a> Adc<'a> {
         unsafe { self.reg.sc1a.read().get_bit(7) }
     }
 
-    pub fn read(&mut self) -> u32 {
-        unsafe { self.reg.ra.read() }
+    pub fn read(&mut self) -> u16 {
+        unsafe { self.reg.ra.read() as u16 }
     }
 }
 
@@ -347,8 +347,8 @@ impl<'a, 'b> AdcDiff<'a, 'b> {
         unsafe { self.reg.sc1a.read().get_bit(7) }
     }
 
-    pub fn read(&mut self) -> i32 {
-        unsafe { self.reg.ra.read() as i32 }
+    pub fn read(&mut self) -> i16 {
+        unsafe { self.reg.ra.read() as i16 }
     }
 }
